feat(app/model_car): improve shell process
