+++ app/model_car/services/throttle/config.h
@@ -2,7 +2,7 @@
  * @Author: andy.chang 
  * @Date: 2024-12-31 10:40:40 
  * @Last Modified by: andy.chang
- * @Last Modified time: 2024-12-31 13:14:52
+ * @Last Modified time: 2024-12-31 13:25:15
  */
 
 #pragma once
